Make RAND_METH opaque
