+++ crypto/include/internal/rand.h
@@ -17,4 +17,13 @@
 
 #include <openssl/rand.h>
 
+struct rand_meth_st {
+    int (*seed) (const void *buf, int num);
+    int (*bytes) (unsigned char *buf, int num);
+    void (*cleanup) (void);
+    int (*add) (const void *buf, int num, double randomness);
+    int (*pseudorand) (unsigned char *buf, int num);
+    int (*status) (void);
+};
+
 void rand_cleanup_int(void);
