+++ include/openssl/rand.h
@@ -19,15 +19,6 @@
 extern "C" {
 #endif
 
-struct rand_meth_st {
-    int (*seed) (const void *buf, int num);
-    int (*bytes) (unsigned char *buf, int num);
-    void (*cleanup) (void);
-    int (*add) (const void *buf, int num, double randomness);
-    int (*pseudorand) (unsigned char *buf, int num);
-    int (*status) (void);
-};
-
 # ifdef BN_DEBUG
 extern int rand_predictable;
 # endif
