+++ test/ecdsatest.c
@@ -27,6 +27,7 @@
 # include <openssl/err.h>
 # include <openssl/rand.h>
 # include "testutil.h"
+# include "internal/rand.h"
 
 static const char rnd_seed[] =
     "string to make the random number generator think it has randomness";
